Fix: duplicated proxy messages (Issue #3203)
Quite simpler solution: does not process the proxy message in the same server as it was sent from. It works fine because the message is sent to the audience in the origin server before the proxy packet is sent.

This results in:

Origin server (lobby), processed as local message:
`[global] Player: message`

Other servers, processed as proxy message:
`[lobby] [global] Player: message`